Moved ColorSpaces to its own file
